Moved common test cases to base class for binary and open xml readers.
